+++ app/src/main/java/com/seungma/infratalk/data/datasource/remote/image/ImageDataSource.kt
@@ -1,4 +1,4 @@
-package com.seungma.infratalk.data.datasource.remote
+package com.seungma.infratalk.data.datasource.remote.image
 
 import com.seungma.infratalk.data.model.request.image.ImagesRequest
 import com.seungma.infratalk.data.model.response.image.ImagesResponse
