+++ app/src/main/java/com/seungma/infratalk/data/datasource/remote/user/FirebaseUserRemoteDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.seungma.infratalk.data.datasource.remote
+package com.seungma.infratalk.data.datasource.remote.user
 
 
 import android.net.Uri
